+++ crates/gmsol/src/exchange/position_cut.rs
@@ -50,7 +50,7 @@ pub struct PositionCutBuilder<'a, C> {
     hint: Option<PositionCutHint>,
     feeds_parser: FeedsParser,
     close: bool,
-    event_buffer_index: u8,
+    event_buffer_index: u16,
     alts: HashMap<Pubkey, Vec<Pubkey>>,
 }
 
@@ -199,7 +199,7 @@ impl<'a, C: Deref<Target = impl Signer> + Clone> PositionCutBuilder<'a, C> {
     }
 
     /// Set event buffer index.
-    pub fn event_buffer_index(&mut self, index: u8) -> &mut Self {
+    pub fn event_buffer_index(&mut self, index: u16) -> &mut Self {
         self.event_buffer_index = index;
         self
     }