+++ programs/gmsol-store/src/states/oracle/feed.rs
@@ -15,9 +15,9 @@ const MAX_FLAGS: usize = 8;
 #[cfg_attr(feature = "debug", derive(Debug))]
 pub struct PriceFeed {
     pub(crate) bump: u8,
-    pub(crate) index: u8,
     pub(crate) provider: u8,
-    padding_0: [u8; 13],
+    pub(crate) index: u16,
+    padding_0: [u8; 12],
     pub(crate) store: Pubkey,
     pub(crate) authority: Pubkey,
     pub(crate) token: Pubkey,
@@ -40,7 +40,7 @@ impl PriceFeed {
     pub(crate) fn init(
         &mut self,
         bump: u8,
-        index: u8,
+        index: u16,
         provider: PriceProviderKind,
         store: &Pubkey,
         authority: &Pubkey,
@@ -48,8 +48,8 @@ impl PriceFeed {
         feed_id: &Pubkey,
     ) -> Result<()> {
         self.bump = bump;
-        self.index = index;
         self.provider = provider.into();
+        self.index = index;
         self.store = *store;
         self.authority = *authority;
         self.token = *token;